% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChangeX.R
\name{ChangeX_F}
\alias{ChangeX_F}
\title{Edit inputs inside FILEX}
\usage{
ChangeX_F(X, section, column_name, new_value)
}
\arguments{
\item{X}{FILEX loaded using readLines, char}

\item{section}{section name, char}

\item{column_name}{header name, char}

\item{new_value}{new value, char of length matching number of lines under the header}
}
\value{
Edited X for further edition or to write using writeLines.
}
\description{
Editing a value in a specified section and under a specified header inside a DSSAT FILEX.
Not vectorised - can currently edit only one input.
}
\examples{
X <- ChangeX_F(X = X, section = "FIELDS", column_name = "WSTA", new_value = "ABCD")
}
